requirements.txt
absl-py==2.2.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.10.1
aiortc==1.9.0
aiosignal==1.4.0
aiosqlite==0.21.0
albucore==0.0.23
albumentations==2.0.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.6.0
APScheduler==3.11.0
astunparse==1.6.3
asyncua==1.1.6
attrs==25.3.0
av==12.3.0
backoff==2.2.1
beautifulsoup4==4.13.5
bleach==6.2.0
blinker==1.9.0
blobfile==3.0.0
boto3==1.35.60
botocore==1.35.99
bs4==0.0.2
cachetools==5.5.2
cadquery==2.5.2
cadquery-ocp==7.7.2
casadi==3.7.0
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
ConfigArgParse==1.7
contourpy==1.3.2
cryptography==45.0.5
ctranslate2==4.6.0
cycler==0.12.1
Cython==3.0.12
dash==3.0.3
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
einops==0.8.0
entmax==1.3
enum34==1.1.10
et_xmlfile==2.0.0
ezdxf==1.4.2
fastapi==0.115.14
faster-whisper==1.1.1
fastjsonschema==2.21.1
filelock==3.16.1
fire==0.6.0
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.7.0
fsspec==2024.9.0
gast==0.6.0
google-crc32c==1.7.1
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
halo==0.0.31
httpcore==1.0.5
httpx==0.28.1
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
importlib_metadata==8.6.1
imutils==0.5.4
inference==0.51.9
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
kaggle==1.7.4.5
keras==3.9.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
llama_models==0.0.36
llama_stack==0.0.36
log-symbols==0.0.14
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.6.3
multimethod==1.12
munch==4.0.0
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.35.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nextcord==2.6.0
nlopt==2.9.1
numpy==2.1.3
nvidia-ml-py==12.575.51
onnxruntime==1.21.1
onvif-zeep-async==2.0.0
open3d==0.19.0
openai==1.97.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
openwakeword==0.6.0
opt_einsum==3.4.0
optree==0.14.1
packaging==24.1
paho-mqtt==1.6.1
pandas==2.2.3
path==17.1.0
pillow==11.3.0
pix2tex==0.1.3
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.6.0
pocketsphinx==5.0.4
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.22.1
prompt_toolkit==3.0.48
propcache==0.3.2
protobuf==5.29.4
psutil==7.0.0
pvporcupine==1.9.5
py-cpuinfo==9.0.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pybase64==1.0.2
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydot==2.0.0
pydub==0.25.1
pyee==13.0.0
PyGetWindow==0.0.9
pyglet==1.5.31
Pygments==2.18.0
PyJWT==2.10.1
pylibsrtp==0.12.0
pylogix==1.0.5
pymodbus==3.8.3
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyperclip==1.9.0
pyquaternion==0.9.9
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
realtimestt==0.3.104
redis==5.0.8
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.8.1
rpds-py==0.24.0
rtree==1.4.0
rubicon-objc==0.5.1
s3transfer==0.10.4
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
shapely==2.0.7
shellingham==1.5.4
simple-pid==2.0.1
simsimd==6.2.1
six==1.16.0
slack_sdk==3.33.5
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.8
SpeechRecognition==3.14.3
spinners==0.0.24
spotipy==2.25.1
starlette==0.46.2
stringzilla==3.11.3
structlog==24.4.0
supervision==0.25.1
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.7.0
timm==0.5.4
tldextract==5.1.3
tokenizers==0.21.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.20.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.48.0
trimesh==4.6.1
twilio==9.3.8
typer==0.16.0
typing-inspect==0.9.0
typing_extensions==4.12.2
typish==1.9.3
tzdata==2024.2
tzlocal==5.3.1
ultralytics==8.3.170
ultralytics-thop==2.0.14
urllib3==2.2.3
wait-for2==0.3.2
wcwidth==0.2.13
webencodings==0.5.1
webrtcvad-wheels==2.0.14
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
wrapt==1.17.2
x-transformers==0.15.0
yarl==1.20.1
youtube-dl==2021.12.17
yt-dlp==2025.7.21
zeep==4.3.1
zipp==3.21.0
zxing-cpp==2.2.0
